+++ src/reactionsystem.jl
@@ -724,24 +724,6 @@ function get_indep_sts(rs::ReactionSystem, remove_conserved = false)
 end
 
 """
-    reactionparams(network)
-
-Given a [`ReactionSystem`](@ref), return a vector of all parameters defined
-within the system and any subsystems that are of type `ReactionSystem`. To get
-the parameters in the system and all subsystems, including non-`ReactionSystem`
-subsystems, use `parameters(network)`.
-
-Notes:
-- Allocates and has to calculate these dynamically by comparison for each reaction.
-"""
-function reactionparams(network)
-    ps = get_ps(network)
-    systems = filter_nonrxsys(network)
-    isempty(systems) && return ps
-    unique([ps; reduce(vcat, map(sys -> species(sys, reactionparams(sys)), systems))])
-end
-
-"""
     numparams(network)
 
 Return the total number of parameters within the given system and all subsystems.
@@ -763,16 +745,6 @@ parameters that appear within the system to their index within
 """
 function paramsmap(network)
     Dict(p => i for (i, p) in enumerate(parameters(network)))
-end
-
-"""
-    reactionparamsmap(network)
-
-Given a [`ReactionSystem`](@ref), return a Dictionary mapping from parameters that
-appear within `Reaction`s to their index within [`reactionparams(network)`](@ref).
-"""
-function reactionparamsmap(network)
-    Dict(p => i for (i, p) in enumerate(reactionparams(network)))
 end
 
 # used in the next function (`reactions(network)`).
@@ -822,21 +794,6 @@ Notes:
 """
 function nonreactions(network)
     equations(network)[(numreactions(network) + 1):end]
-end
-
-"""
-    numreactionparams(network)
-
-Return the total number of parameters within the given [`ReactionSystem`](@ref)
-and subsystems that are `ReactionSystem`s.
-
-Notes
-- If there are no subsystems this will be fast.
-- As this calls [`reactionparams`](@ref), it can be slow and will allocate if
-  there are any subsystems.
-"""
-function numreactionparams(network)
-    length(reactionparams(network))
 end
 
 """